+++ platform-sdk/base-crypto/build.gradle.kts
@@ -14,6 +14,7 @@ testModuleInfo {
     requires("com.swirlds.common")
     requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
+    requires("com.swirlds.logging.test.fixtures")
     requires("org.hiero.base.crypto")
     requires("org.hiero.base.crypto.test.fixtures")
     requires("org.assertj.core")
