+++ platform-sdk/swirlds-common/src/testFixtures/java/module-info.java
@@ -10,7 +10,6 @@ open module com.swirlds.common.test.fixtures {
     exports com.swirlds.common.test.fixtures.set;
     exports com.swirlds.common.test.fixtures.stream;
     exports com.swirlds.common.test.fixtures.platform;
-    exports com.swirlds.common.test.fixtures.logging;
 
     requires transitive com.hedera.pbj.runtime;
     requires transitive com.swirlds.base;
@@ -25,11 +24,11 @@ open module com.swirlds.common.test.fixtures {
     requires transitive org.hiero.consensus.model;
     requires transitive org.hiero.consensus.reconnect;
     requires transitive org.hiero.consensus.utility;
-    requires transitive org.apache.logging.log4j.core;
     requires com.swirlds.config.extensions.test.fixtures;
     requires com.swirlds.logging;
     requires org.hiero.base.utility.test.fixtures;
     requires org.hiero.consensus.metrics;
+    requires org.apache.logging.log4j.core;
     requires org.apache.logging.log4j;
     requires org.junit.jupiter.api;
     requires org.mockito;
