+++ platform-sdk/swirlds-logging/src/testFixtures/java/module-info.java
@@ -8,7 +8,9 @@ open module com.swirlds.logging.test.fixtures {
     requires transitive com.swirlds.config.api;
     requires transitive com.swirlds.config.extensions.test.fixtures;
     requires transitive com.swirlds.logging;
+    requires transitive org.apache.logging.log4j.core;
     requires transitive org.junit.jupiter.api;
     requires com.swirlds.base.test.fixtures;
+    requires org.apache.logging.log4j;
     requires static transitive com.github.spotbugs.annotations;
 }
